feat: rss support dynamic route
