+++ packages/vitepress-plugin-rss/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Updated dependencies
   - @sugarat/theme-shared@0.0.3
+- feat: support rewrites and dynamic routes
 
 ## 0.2.10
