+++ packages/vitepress-plugin-rss/package.json
@@ -39,11 +39,10 @@
   },
   "dependencies": {
     "@sugarat/theme-shared": "workspace:*",
-    "fast-glob": "^3.2.12",
     "feed": "^4.2.2"
   },
   "devDependencies": {
     "vite": "^5",
-    "vitepress": "1.1.3"
+    "vitepress": "^1.3.4"
   }
 }
